[build] Address CVE-2024-43485
Context: https://github.com/advisories/GHSA-8g4q-xg66-9fp4

Consolidates the MSBuild package reference version into a shared prop
and bumps it to the latest to bring in a newer System.Text.Json.

Adds a System.Text.Json reference to Xamarin.AndroidBinderator.csproj to
pull in the latest version of that package with the CVE fix.